+++ resolve_freesound_browser/__init__.py
@@ -1,3 +1,3 @@
 """Resolve Freesound Browser."""
 
-__version__ = "0.3.1"
+__version__ = "0.3.2"